requirements-dev.txt
black
flake8
mypy
boto3-stubs[acm,ssm]
